requirements.txt
# Add this in for tox, comment out for build
--extra-index-url https://download.pytorch.org/whl/cu118
horde_model_reference
pydantic
torch>=2.0.0
xformers>=0.0.19
torchvision
torchaudio
torchdiffeq
torchsde
einops
open-clip-torch
transformers>=4.25.1
safetensors
pytorch_lightning
pynvml
aiohttp
accelerate
pyyaml
pillow
loguru
GitPython
clip-anytorch
diffusers
omegaconf
psutil
typing-extensions
distro
pyarrow
pandas
# face fixer custom nodes
opencv-python
# controlnet custom nodes
opencv-contrib-python
opencv-python
timm>=0.6.12
torchvision
scipy
matplotlib
scikit-image
prettytable>=3.6.0
svglib
reportlab
addict
yapf
mediapipe>=0.9.1.0
fairscale